- replace buttons in HomePage and align content to bottom in slider for 2 of them - moved BluBorderWhiteButton to body package
