+++ lib/widgets/body/ContactUsImageBanner.dart
@@ -1,7 +1,6 @@
-import 'dart:ui';
-
 import 'package:flutter/material.dart';
-import 'package:zengage_learning_platform/constants/app_colors.dart';
+
+import 'BlueBorderWhiteButton.dart';
 
 class ContactUsImageBanner extends StatelessWidget {
   final String imageLink;
@@ -28,53 +27,6 @@ class ContactUsImageBanner extends StatelessWidget {
         ),
         BlueBorderWhiteButton(width: imageWidth, height: imageHeight)
       ],
-    );
-  }
-}
-
-class BlueBorderWhiteButton extends StatelessWidget {
-  BlueBorderWhiteButton({
-    Key key,
-    double width = -1.0,
-    double height = -1.0,
-  }) : super(key: key) {
-    this.width = width;
-    this.height = height;
-  }
-
-  double width;
-  double height;
-
-  @override
-  Widget build(BuildContext context) {
-    if (width == -1.0) {
-      width = MediaQuery.of(context).size.width;
-    }
-    if (height == -1.0) {
-      height = MediaQuery.of(context).size.height;
-    }
-    return Container(
-//      margin: EdgeInsets.only(left: width * 0.1, top: height * 0.35),
-      color: Colors.white,
-      child: OutlineButton(
-        padding: EdgeInsets.fromLTRB(
-            width * 0.1, height * 0.03, width * 0.1, height * 0.03),
-        child: Text(
-          "Contact Us",
-          style: TextStyle(
-              fontSize: (width > 700.0) ? 18.0 : 12.0,
-              color: textGreyColor,
-              fontWeight: FontWeight.bold),
-        ),
-        onPressed: () {
-          print("Contact Us pressed");
-        },
-        borderSide: BorderSide(
-          color: Colors.lightBlue, //Color of the border
-          style: BorderStyle.solid, //Style of the border
-          width: 4, //width of the border
-        ),
-      ),
     );
   }
 }
